Moved to compiler
